chore: relocate files and clean up unused codes
